Rename FramedAutoGenerator<T> -> AutoGenerator<T>
