+++ Templates/Rulesets/ruleset-scrolling-example/osu.Game.Rulesets.Pippidon/Replays/PippidonAutoGenerator.cs
@@ -9,7 +9,7 @@ using osu.Game.Rulesets.Replays;
 
 namespace osu.Game.Rulesets.Pippidon.Replays
 {
-    public class PippidonAutoGenerator : FramedAutoGenerator<PippidonReplayFrame>
+    public class PippidonAutoGenerator : AutoGenerator<PippidonReplayFrame>
     {
         public new Beatmap<PippidonHitObject> Beatmap => (Beatmap<PippidonHitObject>)base.Beatmap;
