+++ osu.Game/Rulesets/Replays/AutoGenerator.cs
@@ -1,6 +1,9 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Collections.Generic;
+using System.Linq;
+using JetBrains.Annotations;
 using osu.Game.Beatmaps;
 using osu.Game.Replays;
 using osu.Game.Rulesets.Objects;
@@ -36,5 +39,38 @@ namespace osu.Game.Rulesets.Replays
 
             return Beatmap.HitObjects[currentIndex + 1];
         }
+    }
+
+    public abstract class AutoGenerator<TFrame> : AutoGenerator
+        where TFrame : ReplayFrame
+    {
+        /// <summary>
+        /// The replay frames of the autoplay.
+        /// </summary>
+        protected readonly List<TFrame> Frames = new List<TFrame>();
+
+        [CanBeNull]
+        protected TFrame LastFrame => Frames.Count == 0 ? null : Frames[^1];
+
+        protected AutoGenerator(IBeatmap beatmap)
+            : base(beatmap)
+        {
+        }
+
+        public sealed override Replay Generate()
+        {
+            Frames.Clear();
+            GenerateFrames();
+
+            return new Replay
+            {
+                Frames = Frames.OrderBy(frame => frame.Time).Cast<ReplayFrame>().ToList()
+            };
+        }
+
+        /// <summary>
+        /// Generate the replay frames of the autoplay and populate <see cref="Frames"/>.
+        /// </summary>
+        protected abstract void GenerateFrames();
     }
 }
